Renamed model/view files
