+++ lib/pages/stream/stream_model.dart
