+++ lib/pages/stream/stream_view.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:endoscopy_ai/shared/widget/screenshot_feed.dart';
 import 'package:endoscopy_ai/shared/widget/spacing.dart';
 import 'package:provider/provider.dart';
-import 'package:endoscopy_ai/pages/stream/stream_page_model.dart';
+import 'package:endoscopy_ai/pages/stream/stream_model.dart';
 import 'package:endoscopy_ai/shared/widget/buttons.dart';
 
 //  Логика, содержащая логику, связанную с UI
