+++ lib/shared/widget/gallery_of_screenshot.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:endoscopy_ai/pages/file_video/file_video_page_model.dart';
+import 'package:endoscopy_ai/pages/file_video/file_video_model.dart';
 import 'package:endoscopy_ai/shared/widget/screenshot_preview.dart';
 
 class GalleryOfScreenshot extends StatelessWidget {
